passt: Move passt related code to its own package
