+++ controllers/handlers/kubevirt.go
@@ -28,6 +28,7 @@ import (
 
 	hcov1beta1 "github.com/kubevirt/hyperconverged-cluster-operator/api/v1beta1"
 	"github.com/kubevirt/hyperconverged-cluster-operator/controllers/common"
+	"github.com/kubevirt/hyperconverged-cluster-operator/controllers/handlers/passt"
 	"github.com/kubevirt/hyperconverged-cluster-operator/controllers/operands"
 	"github.com/kubevirt/hyperconverged-cluster-operator/pkg/nodeinfo"
 	"github.com/kubevirt/hyperconverged-cluster-operator/pkg/patch"
@@ -53,28 +54,12 @@ const (
 
 const (
 	primaryUDNNetworkBindingName = "l2bridge"
-
-	deployPasstNetworkBindingAnnotation = "deployPasstNetworkBinding"
-
-	passtBindingName                = "passt"
-	passtNetworkBindingNADName      = "primary-udn-kubevirt-binding"
-	passtNetworkBindingNADNamespace = "default"
-	networkAttachmentDefinition     = passtNetworkBindingNADNamespace + "/" + passtNetworkBindingNADName
-
-	bindingComputeMemoryOverhead = "250Mi"
 )
 
 const kvPriorityClass = "kubevirt-cluster-critical"
 
 var (
 	useKVMEmulation = false
-)
-
-var passtResourceMemory = resource.MustParse(bindingComputeMemoryOverhead)
-
-var (
-	passtImage     string
-	passtImageOnce sync.Once
 )
 
 func init() {
@@ -554,8 +539,8 @@ func getNetworkBindings(hcoNetworkBindings map[string]kubevirtcorev1.InterfaceBi
 
 	networkBindings[primaryUDNNetworkBindingName] = primaryUserDefinedNetworkBinding()
 
-	if hcoAnnotations[deployPasstNetworkBindingAnnotation] == "true" {
-		networkBindings[passtBindingName] = passtNetworkBinding(getPasstImage())
+	if hcoAnnotations[passt.DeployPasstNetworkBindingAnnotation] == "true" {
+		networkBindings[passt.BindingName] = passt.NetworkBinding()
 	}
 	return networkBindings
 }
@@ -875,7 +860,7 @@ func getFeatureGateChecks(featureGates *hcov1beta1.HyperConvergedFeatureGates, a
 		fgs = append(fgs, kvAlignCPUs)
 	}
 
-	if annotations[deployPasstNetworkBindingAnnotation] == "true" {
+	if annotations[passt.DeployPasstNetworkBindingAnnotation] == "true" {
 		fgs = append(fgs, kvPasstIPStackMigration)
 	}
 
@@ -1045,24 +1030,4 @@ func getLabelPatch(dest, src map[string]string) ([]byte, error) {
 	}
 
 	return json.Marshal(patches)
-}
-
-func passtNetworkBinding(sidecarImage string) kubevirtcorev1.InterfaceBindingPlugin {
-	return kubevirtcorev1.InterfaceBindingPlugin{
-		NetworkAttachmentDefinition: networkAttachmentDefinition,
-		SidecarImage:                sidecarImage,
-		Migration:                   &kubevirtcorev1.InterfaceBindingMigration{},
-		ComputeResourceOverhead: &kubevirtcorev1.ResourceRequirementsWithoutClaims{
-			Requests: corev1.ResourceList{
-				corev1.ResourceMemory: passtResourceMemory,
-			},
-		},
-	}
-}
-
-func getPasstImage() string {
-	passtImageOnce.Do(func() {
-		passtImage = os.Getenv(hcoutil.PasstImageEnvV)
-	})
-	return passtImage
 }
